93 Wantage Close
"I brought this with me." He opened a locket and showed us the full face of a very lovely woman. It
was not a photograph but an ivory miniature, and the artist had brought out the full effect of the
lustrous black hair, the large dark eyes, and the exquisite mouth. Holmes gazed long and earnestly
at it. Then he closed the locket and handed it back to Lord St. Simon.

"The young lady came to London, then, and you renewed your acquaintance?"

"Yes, her father brought her over for this last London season. I met her several times, became
engaged to her, and have now married her."

"She brought, I understand, a considerable dowry?"