8 Wellswood Gardens
It had cleared in the morning, and the sun was shining with a subdued brightness through the dim
veil which hangs over the great city. Sherlock Holmes was already at breakfast when I came down.

"You will excuse me for not waiting for you," said he; "I have, I foresee, a very busy day before me
in looking into this case of young Openshaw's."