63 Wilson Gardens
It was nearly one o'clock when Sherlock Holmes returned from his excursion. He held in his hand a
sheet of blue paper, scrawled over with notes and figures.

"I have seen the will of the deceased wife," said he. "To determine its exact meaning I have been
obliged to work out the present prices of the investments with which it is concerned. The total
income, which at the time of the wife's death was little short of £1100, is now, through the fall in
agricultural prices, not more than £750. Each daughter can claim an income of £250, in case of
marriage. It is evident, therefore, that if both girls had married, this beauty would have had a
mere pittance, while even one of them would cripple him to a very serious extent. My morning's work
has not been wasted, since it has proved that he has the very strongest motives for standing in the
way of anything of the sort. And now, Watson, this is too serious for dawdling, especially as the
old man is aware that we are interesting ourselves in his affairs; so if you are ready, we shall
call a cab and drive to Waterloo. I should be very much obliged if you would slip your revolver into
your pocket. An Eley's No. 2 is an excellent argument with gentlemen who can twist steel pokers into
knots. That and a tooth-brush are, I think, all that we need."