85 Holmefield Avenue
"I would pay ten."

"That would be unnecessary. Three thousand will cover the matter. And there is a little reward, I
fancy. Have you your check-book?  Here is a pen. Better make it out for £4000."

With a dazed face the banker made out the required check. Holmes walked over to his desk, took out a
little triangular piece of gold with three gems in it, and threw it down upon the table.

With a shriek of joy our client clutched it up.

"You have it!" he gasped. "I am saved! I am saved!"