33 Sunningdale Drive
I could see that Holmes was favourably impressed by the manner and speech of his new client. He
looked her over in his searching fashion, and then composed himself, with his lids drooping and his
finger-tips together, to listen to her story.

"I have been a governess for five years," said she, "in the family of Colonel Spence Munro, but two
months ago the colonel received an appointment at Halifax, in Nova Scotia, and took his children
over to America with him, so that I found myself without a situation. I advertised, and I answered
advertisements, but without success. At last the little money which I had saved began to run short,
and I was at my wit's end as to what I should do.

"There is a well-known agency for governesses in the West End called Westaway's, and there I used to
call about once a week in order to see whether anything had turned up which might suit me. Westaway
was the name of the founder of the business, but it is really managed by Miss Stoper. She sits in
her own little office, and the ladies who are seeking employment wait in an anteroom, and are then
shown in one by one, when she consults her ledgers and sees whether she has anything which would
suit them.