140 Sunningdale Drive
"Yesterday, however, just as I was thinking of leaving the office, my clerk entered to say there was
a gentleman waiting who wished to see me upon business. He brought up a card, too, with the name of
'Colonel Lysander Stark' engraved upon it. Close at his heels came the colonel himself, a man rather
over the middle size, but of an exceeding thinness. I do not think that I have ever seen so thin a
man. His whole face sharpened away into nose and chin, and the skin of his cheeks was drawn quite
tense over his outstanding bones. Yet this emaciation seemed to be his natural habit, and due to no
disease, for his eye was bright, his step brisk, and his bearing assured. He was plainly but neatly
dressed, and his age, I should judge, would be nearer forty than thirty.

"'Mr. Hatherley?' said he, with something of a German accent. 'You have been recommended to me, Mr.
Hatherley, as being a man who is not only proficient in his profession but is also discreet and
capable of preserving a secret.'

"I bowed, feeling as flattered as any young man would at such an address. 'May I ask who it was who
gave me so good a character?'

"'Well, perhaps it is better that I should not tell you that just at this moment. I have it from the
same source that you are both an orphan and a bachelor and are residing alone in London.'

"'That is quite correct,' I answered; 'but you will excuse me if I say that I cannot see how all
this bears upon my professional qualifications. I understand that it was on a professional matter
that you wished to speak to me?'

"'Undoubtedly so. But you will find that all I say is really to the point. I have a professional
commission for you, but absolute secrecy is quite essential--absolute secrecy, you understand, and
of course we may expect that more from a man who is alone than from one who lives in the bosom of
his family.'