35 Wilson Gardens
"If I may give an opinion," remarked the strange gentleman, "we've had just a little too much
secrecy over this business already. For my part, I should like all Europe and America to hear the
rights of it." He was a small, wiry, sunburnt man, clean-shaven, with a sharp face and alert manner.

"Then I'll tell our story right away," said the lady. "Frank here and I met in '84, in McQuire's
camp, near the Rockies, where pa was working a claim. We were engaged to each other, Frank and I;
but then one day father struck a rich pocket and made a pile, while poor Frank here had a claim that
petered out and came to nothing. The richer pa grew the poorer was Frank; so at last pa wouldn't
hear of our engagement lasting any longer, and he took me away to 'Frisco. Frank wouldn't throw up
his hand, though; so he followed me there, and he saw me without pa knowing anything about it. It
would only have made him mad to know, so we just fixed it all up for ourselves. Frank said that he
would go and make his pile, too, and never come back to claim me until he had as much as pa. So then
I promised to wait for him to the end of time and pledged myself not to marry anyone else while he
lived. 'Why shouldn't we be married right away, then,' said he, 'and then I will feel sure of you;
and I won't claim to be your husband until I come back?' Well, we talked it over, and he had fixed
it all up so nicely, with a clergyman all ready in waiting, that we just did it right there; and
then Frank went off to seek his fortune, and I went back to pa.

"The next I heard of Frank was that he was in Montana, and then he went prospecting in Arizona, and
then I heard of him from New Mexico. After that came a long newspaper story about how a miners' camp
had been attacked by Apache Indians, and there was my Frank's name among the killed. I fainted dead
away, and I was very sick for months after. Pa thought I had a decline and took me to half the
doctors in 'Frisco. Not a word of news came for a year and more, so that I never doubted that Frank
was really dead. Then Lord St. Simon came to 'Frisco, and we came to London, and a marriage was
arranged, and pa was very pleased, but I felt all the time that no man on this earth would ever take
the place in my heart that had been given to my poor Frank.