71 Wantage Close
It was about ten minutes before we regained our cab and drove back into Ross, Holmes still carrying
with him the stone which he had picked up in the wood.

"This may interest you, Lestrade," he remarked, holding it out.  "The murder was done with it."

"I see no marks."