179 Holmefield Avenue
"I promise," said Holmes.

"And I."

"You will excuse this mask," continued our strange visitor. "The august person who employs me wishes
his agent to be unknown to you, and I may confess at once that the title by which I have just called
myself is not exactly my own."

"I was aware of it," said Holmes dryly.

"The circumstances are of great delicacy, and every precaution has to be taken to quench what might
grow to be an immense scandal and seriously compromise one of the reigning families of Europe. To
speak plainly, the matter implicates the great House of Ormstein, hereditary kings of Bohemia."

"I was also aware of that," murmured Holmes, settling himself down in his armchair and closing his
eyes.