55 Bowler Street
"Yes. That is what Mr. Lestrade, of Scotland Yard, looks upon as so serious. It is thought that
Flora decoyed my wife out and laid some terrible trap for her."

"Well, it is a possible supposition."

"You think so, too?"

"I did not say a probable one. But you do not yourself look upon this as likely?"

"I do not think Flora would hurt a fly."

"Still, jealousy is a strange transformer of characters. Pray what is your own theory as to what
took place?"