59 Ventnor Terrace
"And when I raise my hand--so--you will throw into the room what I give you to throw, and will, at
the same time, raise the cry of fire. You quite follow me?"

"Entirely."

"It is nothing very formidable," he said, taking a long cigar-shaped roll from his pocket. "It is an
ordinary plumber's smoke-rocket, fitted with a cap at either end to make it self-lighting. Your task
is confined to that. When you raise your cry of fire, it will be taken up by quite a number of
people. You may then walk to the end of the street, and I will rejoin you in ten minutes. I hope
that I have made myself clear?"