73 Bowler Street
On glancing over my notes of the seventy odd cases in which I have during the last eight years
studied the methods of my friend Sherlock Holmes, I find many tragic, some comic, a large number
merely strange, but none commonplace; for, working as he did rather for the love of his art than for
the acquirement of wealth, he refused to associate himself with any investigation which did not tend
towards the unusual, and even the fantastic. Of all these varied cases, however, I cannot recall any
which presented more singular features than that which was associated with the well-known Surrey
family of the Roylotts of Stoke Moran. The events in question occurred in the early days of my
association with Holmes, when we were sharing rooms as bachelors in Baker Street. It is possible
that I might have placed them upon record before, but a promise of secrecy was made at the time,
from which I have only been freed during the last month by the untimely death of the lady to whom
the pledge was given. It is perhaps as well that the facts should now come to light, for I have
reasons to know that there are widespread rumours as to the death of Dr. Grimesby Roylott which tend
to make the matter even more terrible than the truth.

It was early in April in the year '83 that I woke one morning to find Sherlock Holmes standing,
fully dressed, by the side of my bed. He was a late riser, as a rule, and as the clock on the
mantelpiece showed me that it was only a quarter-past seven, I blinked up at him in some surprise,
and perhaps just a little resentment, for I was myself regular in my habits.

"Very sorry to knock you up, Watson," said he, "but it's the common lot this morning. Mrs. Hudson
has been knocked up, she retorted upon me, and I on you."

"What is it, then--a fire?"