155 Holmefield Avenue
"Your boy, Arthur, went to bed after his interview with you but he slept badly on account of his
uneasiness about his club debts. In the middle of the night he heard a soft tread pass his door, so
he rose and, looking out, was surprised to see his cousin walking very stealthily along the passage
until she disappeared into your dressing-room. Petrified with astonishment, the lad slipped on some
clothes and waited there in the dark to see what would come of this strange affair. Presently she
emerged from the room again, and in the light of the passage-lamp your son saw that she carried the
precious coronet in her hands. She passed down the stairs, and he, thrilling with horror, ran along
and slipped behind the curtain near your door, whence he could see what passed in the hall beneath.
He saw her stealthily open the window, hand out the coronet to someone in the gloom, and then
closing it once more hurry back to her room, passing quite close to where he stood hid behind the
curtain.

"As long as she was on the scene he could not take any action without a horrible exposure of the
woman whom he loved. But the instant that she was gone he realised how crushing a misfortune this
would be for you, and how all-important it was to set it right. He rushed down, just as he was, in
his bare feet, opened the window, sprang out into the snow, and ran down the lane, where he could
see a dark figure in the moonlight. Sir George Burnwell tried to get away, but Arthur caught him,
and there was a struggle between them, your lad tugging at one side of the coronet, and his opponent
at the other. In the scuffle, your son struck Sir George and cut him over the eye. Then something
suddenly snapped, and your son, finding that he had the coronet in his hands, rushed back, closed
the window, ascended to your room, and had just observed that the coronet had been twisted in the
struggle and was endeavouring to straighten it when you appeared upon the scene."

"Is it possible?" gasped the banker.

"You then roused his anger by calling him names at a moment when he felt that he had deserved your
warmest thanks. He could not explain the true state of affairs without betraying one who certainly
deserved little enough consideration at his hands. He took the more chivalrous view, however, and
preserved her secret."

"And that was why she shrieked and fainted when she saw the coronet," cried Mr. Holder. "Oh, my God!
what a blind fool I have been! And his asking to be allowed to go out for five minutes! The dear
fellow wanted to see if the missing piece were at the scene of the struggle. How cruelly I have
misjudged him!"

"When I arrived at the house," continued Holmes, "I at once went very carefully round it to observe
if there were any traces in the snow which might help me. I knew that none had fallen since the
evening before, and also that there had been a strong frost to preserve impressions. I passed along
the tradesmen's path, but found it all trampled down and indistinguishable. Just beyond it, however,
at the far side of the kitchen door, a woman had stood and talked with a man, whose round
impressions on one side showed that he had a wooden leg. I could even tell that they had been
disturbed, for the woman had run back swiftly to the door, as was shown by the deep toe and light
heel marks, while Wooden-leg had waited a little, and then had gone away. I thought at the time that
this might be the maid and her sweetheart, of whom you had already spoken to me, and inquiry showed
it was so. I passed round the garden without seeing anything more than random tracks, which I took
to be the police; but when I got into the stable lane a very long and complex story was written in
the snow in front of me.