17 Ventnor Terrace
"It is a curious thing," remarked Holmes, "that a typewriter has really quite as much individuality
as a man's handwriting. Unless they are quite new, no two of them write exactly alike. Some letters
get more worn than others, and some wear only on one side. Now, you remark in this note of yours,
Mr. Windibank, that in every case there is some little slurring over of the 'e,' and a slight defect
in the tail of the 'r.' There are fourteen other characteristics, but those are the more obvious."

"We do all our correspondence with this machine at the office, and no doubt it is a little worn,"
our visitor answered, glancing keenly at Holmes with his bright little eyes.