107 Wilson Gardens
"Pray compose yourself, sir," said Holmes, "and let me have a clear account of who you are and what
it is that has befallen you."

"My name," answered our visitor, "is probably familiar to your ears. I am Alexander Holder, of the
banking firm of Holder & Stevenson, of Threadneedle Street."

The name was indeed well known to us as belonging to the senior partner in the second largest
private banking concern in the City of London. What could have happened, then, to bring one of the
foremost citizens of London to this most pitiable pass? We waited, all curiosity, until with another
effort he braced himself to tell his story.

"I feel that time is of value," said he; "that is why I hastened here when the police inspector
suggested that I should secure your co-operation. I came to Baker Street by the Underground and
hurried from there on foot, for the cabs go slowly through this snow. That is why I was so out of
breath, for I am a man who takes very little exercise. I feel better now, and I will put the facts
before you as shortly and yet as clearly as I can.

"It is, of course, well known to you that in a successful banking business as much depends upon our
being able to find remunerative investments for our funds as upon our increasing our connection and
the number of our depositors. One of our most lucrative means of laying out money is in the shape of
loans, where the security is unimpeachable. We have done a good deal in this direction during the
last few years, and there are many noble families to whom we have advanced large sums upon the
security of their pictures, libraries, or plate.

"Yesterday morning I was seated in my office at the bank when a card was brought in to me by one of
the clerks. I started when I saw the name, for it was that of none other than--well, perhaps even to
you I had better say no more than that it was a name which is a household word all over the earth--
one of the highest, noblest, most exalted names in England. I was overwhelmed by the honour and
attempted, when he entered, to say so, but he plunged at once into business with the air of a man
who wishes to hurry quickly through a disagreeable task.