147 Moor End
"That will do, Mr. Wilson. I shall be happy to give you an opinion upon the subject in the course of
a day or two. To-day is Saturday, and I hope that by Monday we may come to a conclusion."

"Well, Watson," said Holmes when our visitor had left us, "what do you make of it all?"

"I make nothing of it," I answered frankly. "It is a most mysterious business."

"As a rule," said Holmes, "the more bizarre a thing is the less mysterious it proves to be. It is
your commonplace, featureless crimes which are really puzzling, just as a commonplace face is the
most difficult to identify. But I must be prompt over this matter."

"What are you going to do, then?" I asked.