83 Badgers Dene
"I had come to these conclusions before ever I had entered his room. An inspection of his chair
showed me that he had been in the habit of standing on it, which of course would be necessary in
order that he should reach the ventilator. The sight of the safe, the saucer of milk, and the loop
of whipcord were enough to finally dispel any doubts which may have remained. The metallic clang
heard by Miss Stoner was obviously caused by her stepfather hastily closing the door of his safe
upon its terrible occupant. Having once made up my mind, you know the steps which I took in order to
put the matter to the proof. I heard the creature hiss as I have no doubt that you did also, and I
instantly lit the light and attacked it."

"With the result of driving it through the ventilator."

"And also with the result of causing it to turn upon its master at the other side. Some of the blows
of my cane came home and roused its snakish temper, so that it flew upon the first person it saw. In
this way I am no doubt indirectly responsible for Dr. Grimesby Roylott's death, and I cannot say
that it is likely to weigh very heavily upon my conscience."