105 Balcombe Close
"He knows it already. When I had cleared it all up I had an interview with him, and finding that he
would not tell me the story, I told it to him, on which he had to confess that I was right and to
add the very few details which were not yet quite clear to me. Your news of this morning, however,
may open his lips."

"For heaven's sake, tell me, then, what is this extraordinary mystery!"