16 Wantage Close
"Yes, my correspondence has certainly the charm of variety," he answered, smiling, "and the humbler
are usually the more interesting. This looks like one of those unwelcome social summonses which call
upon a man either to be bored or to lie."

He broke the seal and glanced over the contents.