94 Harvey Street
"'Oh,' says she, 'we've set yours aside for you--Jem's bird, we call it. It's the big white one over
yonder. There's twenty-six of them, which makes one for you, and one for us, and two dozen for the
market.'

"'Thank you, Maggie,' says I; 'but if it is all the same to you, I'd rather have that one I was
handling just now.'

"'The other is a good three pound heavier,' said she, 'and we fattened it expressly for you.'

"'Never mind. I'll have the other, and I'll take it now,' said I.

"'Oh, just as you like,' said she, a little huffed. 'Which is it you want, then?'

"'That white one with the barred tail, right in the middle of the flock.'