8 Valentia Road
"Convinced that something was amiss with him, she rushed down the steps--for the house was none
other than the opium den in which you found me to-night--and running through the front room she
attempted to ascend the stairs which led to the first floor. At the foot of the stairs, however, she
met this Lascar scoundrel of whom I have spoken, who thrust her back and, aided by a Dane, who acts
as assistant there, pushed her out into the street. Filled with the most maddening doubts and fears,
she rushed down the lane and, by rare good-fortune, met in Fresno Street a number of constables with
an inspector, all on their way to their beat. The inspector and two men accompanied her back, and in
spite of the continued resistance of the proprietor, they made their way to the room in which Mr.
St. Clair had last been seen. There was no sign of him there. In fact, in the whole of that floor
there was no one to be found save a crippled wretch of hideous aspect, who, it seems, made his home
there. Both he and the Lascar stoutly swore that no one else had been in the front room during the
afternoon. So determined was their denial that the inspector was staggered, and had almost come to
believe that Mrs. St. Clair had been deluded when, with a cry, she sprang at a small deal box which
lay upon the table and tore the lid from it. Out there fell a cascade of children's bricks. It was
the toy which he had promised to bring home.

"This discovery, and the evident confusion which the cripple showed, made the inspector realise that
the matter was serious. The rooms were carefully examined, and results all pointed to an abominable
crime. The front room was plainly furnished as a sitting-room and led into a small bedroom, which
looked out upon the back of one of the wharves. Between the wharf and the bedroom window is a narrow
strip, which is dry at low tide but is covered at high tide with at least four and a half feet of
water. The bedroom window was a broad one and opened from below. On examination traces of blood were
to be seen upon the windowsill, and several scattered drops were visible upon the wooden floor of
the bedroom. Thrust away behind a curtain in the front room were all the clothes of Mr. Neville St.
Clair, with the exception of his coat. His boots, his socks, his hat, and his watch--all were there.
There were no signs of violence upon any of these garments, and there were no other traces of Mr.
Neville St. Clair. Out of the window he must apparently have gone for no other exit could be
discovered, and the ominous bloodstains upon the sill gave little promise that he could save himself
by swimming, for the tide was at its very highest at the moment of the tragedy.

"And now as to the villains who seemed to be immediately implicated in the matter. The Lascar was
known to be a man of the vilest antecedents, but as, by Mrs. St. Clair's story, he was known to have
been at the foot of the stair within a very few seconds of her husband's appearance at the window,
he could hardly have been more than an accessory to the crime. His defence was one of absolute
ignorance, and he protested that he had no knowledge as to the doings of Hugh Boone, his lodger, and
that he could not account in any way for the presence of the missing gentleman's clothes.

"So much for the Lascar manager. Now for the sinister cripple who lives upon the second floor of the
opium den, and who was certainly the last human being whose eyes rested upon Neville St. Clair. His
name is Hugh Boone, and his hideous face is one which is familiar to every man who goes much to the
City. He is a professional beggar, though in order to avoid the police regulations he pretends to a
small trade in wax vestas. Some little distance down Threadneedle Street, upon the left-hand side,
there is, as you may have remarked, a small angle in the wall. Here it is that this creature takes
his daily seat, cross-legged with his tiny stock of matches on his lap, and as he is a piteous
spectacle a small rain of charity descends into the greasy leather cap which lies upon the pavement
beside him. I have watched the fellow more than once before ever I thought of making his
professional acquaintance, and I have been surprised at the harvest which he has reaped in a short
time. His appearance, you see, is so remarkable that no one can pass him without observing him. A
shock of orange hair, a pale face disfigured by a horrible scar, which, by its contraction, has
turned up the outer edge of his upper lip, a bulldog chin, and a pair of very penetrating dark eyes,
which present a singular contrast to the colour of his hair, all mark him out from amid the common
crowd of mendicants and so, too, does his wit, for he is ever ready with a reply to any piece of
chaff which may be thrown at him by the passers-by. This is the man whom we now learn to have been
the lodger at the opium den, and to have been the last man to see the gentleman of whom we are in
quest."

"But a cripple!" said I. "What could he have done single-handed against a man in the prime of life?"

"He is a cripple in the sense that he walks with a limp; but in other respects he appears to be a
powerful and well-nurtured man. Surely your medical experience would tell you, Watson, that weakness
in one limb is often compensated for by exceptional strength in the others."