50 Sunningdale Drive
He held out his snuffbox of old gold, with a great amethyst in the centre of the lid. Its splendour
was in such contrast to his homely ways and simple life that I could not help commenting upon it.

"Ah," said he, "I forgot that I had not seen you for some weeks. It is a little souvenir from the
King of Bohemia in return for my assistance in the case of the Irene Adler papers."

"And the ring?" I asked, glancing at a remarkable brilliant which sparkled upon his finger.